requirements/local.txt
-r base.txt

# ==========================================
# Development Tools
# ==========================================
django-debug-toolbar>=4.3.0  # SQL 쿼리 속도 등 분석 도구
django-browser-reload>=1.12.0  # 코드를 수정하면 브라우저가 자동 새로고침됨 (Tailwind와 찰떡궁합)

# ==========================================
# Code Quality
# ==========================================
black>=24.2.0  # 코드 예쁘게 정렬해주는 포맷터